update typing and constants
