+++ src/components/HomePage/PlayRadioButton.tsx
@@ -9,12 +9,12 @@ import PlayIcon from '../../../public/icons/play-arrow.svg';
 import Button from '../dls/Button/Button';
 import Link from '../dls/Link/Link';
 import Spinner from '../dls/Spinner/Spinner';
+import { getRandomCuratedStationId } from '../Radio/curatedStations';
 
 import styles from './PlayRadioButton.module.scss';
 
 import { logEvent } from 'src/utils/eventLogger';
 import { AudioPlayerMachineContext } from 'src/xstate/AudioPlayerMachineContext';
-import { getRandomCuratedStationId } from 'src/xstate/Radio/curatedStations';
 import StationType from 'src/xstate/Radio/types/StationType';
 import useCurrentStationInfo from 'src/xstate/Radio/useCurrentStationInfo';
