--- src/xstate/Radio/types/StationInfo.ts
@@ -1,6 +1,0 @@
-type StationInfo = {
-  title: string;
-  description: string;
-};
-
-export default StationInfo;
