+++ src/xstate/Radio/useCurrentStationInfo.ts
@@ -3,9 +3,8 @@ import useSWRImmutable from 'swr/immutable';
 
 import RadioContext from '../actors/radio/types/RadioContext';
 
-import curatedStations from './curatedStations';
-
 import { getReciterData } from 'src/api';
+import curatedStations from 'src/components/Radio/curatedStations';
 import { StationInfo, StationType } from 'src/components/Radio/types';
 import { makeReciterUrl } from 'src/utils/apiPaths';
