--- src/xstate/actors/repeatMachine/types/Segment.ts
@@ -1,3 +1,0 @@
-type Segment = [number, number, number];
-
-export default Segment;
